+++ src/layouts/header/iconMenu/IconMenu.tsx
@@ -1,18 +1,18 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faMagnifyingGlass } from "@fortawesome/free-solid-svg-icons";
 import { faCommentDots } from "@fortawesome/free-regular-svg-icons";
-import { avatar, iconmenu } from "../../../styles/header.css";
+import { avatar, iconmenu } from "./IconMenu.css";
 import WriteButton from "./writeButton/WriteButton";
 
 const IconMenu: React.FC = () => {
-  return (
-    <div className={iconmenu}>
-      <WriteButton />
-      <FontAwesomeIcon icon={faMagnifyingGlass} />
-      <FontAwesomeIcon icon={faCommentDots} />
-      <div className={avatar}></div>
-    </div>
-  );
+    return (
+        <div className={iconmenu}>
+            <WriteButton />
+            <FontAwesomeIcon icon={faMagnifyingGlass} />
+            <FontAwesomeIcon icon={faCommentDots} />
+            <div className={avatar}></div>
+        </div>
+    );
 };
 
 export default IconMenu;
