+++ src/layouts/header/iconMenu/writeButton/WriteButton.tsx
@@ -1,14 +1,14 @@
 import React from "react";
-import { writeButton } from "../../../../styles/header.css";
+import { writeButton } from "./WriteButton.css";
 import { useNavigate } from "react-router-dom";
 
 const WriteButton: React.FC = () => {
-  const navigate = useNavigate();
-  return (
-    <button className={writeButton} onClick={() => navigate("/posting")}>
-      새 글 작성
-    </button>
-  );
+    const navigate = useNavigate();
+    return (
+        <button className={writeButton} onClick={() => navigate("/posting")}>
+            새 글 작성
+        </button>
+    );
 };
 
 export default WriteButton;
