+++ src/pages/posting/Posting.css.ts
